requirements.txt
langgraph
langchain-google-genai
langchain-core
python-dotenv
firecrawl-py
playwright